% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrisdata.r
\docType{data}
\name{AinsleyGenes}
\alias{AinsleyGenes}
\title{Ainsley adult hippocampal CA1 dendritic genes}
\format{character vector of Entrez gene symbols}
\source{
Ainsley, J.A., Drane, L., Jacobs, J., Kittelberger, K.A.,
Reijmers, L.G., 2014.
Functionally diverse dendritic mRNAs rapidly associate with
ribosomes following a novel experience. Nature Communications 5, 4510.
doi:10.1038/ncomms5510
}
\usage{
AinsleyGenes
}
\description{
Ainsley adult hippocampal CA1 dendritic genes
}
\details{
Genes derived from Ainsley et al., 2014 mouse adult hippocampal CA1
dendrites.

Gene symbols were updated using Bioconductor package
\code{org.Mm.eg.db}, using the reported `gene_name` or
`Entrez Gene ID` from the supplemental data. This step was performed
in order to ensure each reference dataset, including the Farris et al.
data, were using the same version of the gene symbol annotations.
}
\keyword{datasets}
