% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrisdata.r
\docType{data}
\name{CajigasGenes}
\alias{CajigasGenes}
\title{Cajigas rat hippocampal CA1 dendritic genes}
\format{character vector of Entrez gene symbols}
\source{
Cajigas, I.J., Tushev, G., Will, T.J., tom Dieck, S., Fuerst, N.,
Schuman, E.M., 2012.
The local transcriptome in the synaptic neuropil revealed by
deep sequencing and high-resolution imag-ing.
Neuron 74, 453–466.
doi:10.1016/j.neuron.2012.02.036
}
\usage{
CajigasGenes
}
\description{
Cajigas rat hippocampal CA1 dendritic genes
}
\details{
Genes derived from Cajigas et al., 2012 rat hippocampal CA1
dendrites.

Gene symbols were updated using Bioconductor package
\code{org.Rn.eg.db}, using the reported \code{gene_name} or
\code{Entrez Gene ID} from the supplemental data. This step was performed
in order to ensure each reference dataset, including the Farris et al.
data, were using the same version of the gene symbol annotations.
}
\keyword{datasets}
