% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrisdata.r
\docType{data}
\name{colorSub}
\alias{colorSub}
\title{Group colors used for Farris data}
\format{character vector of R color names, named by the
experiment factors, sample groups, and boolean values TRUE
and FALSE.}
\usage{
colorSub
}
\description{
Group colors used for Farris data
}
\details{
The \code{colorSub} object is simple a vector of colors,
named by the experimental factors, the sample groups,
and boolean TRUE and FALSE.
}
\examples{
data(colorSub);
print(colorSub);

# Split into a list, just for visualization
colorSubType <- ifelse(grepl("_", names(colorSub)),
   "Group",
   "Factor");
colorSubL <- split(colorSub, colorSubType);
if (require(jamba)) {
   showColors(colorSubL);
}
}
\concept{Farris et al colors}
\keyword{datasets}
