% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrisdata-enrich.r
\docType{data}
\name{EnrichmentDFL}
\alias{EnrichmentDFL}
\title{Farris et al gene set enrichment data.frames}
\format{
\code{list} of \code{data.frame} objects each with these columns:
\itemize{
\item \code{"Comparison"} - \code{character} string indicating the statistical
comparison, and direction of change.
\item \code{"P-value"} - \code{numeric} adjusted enrichment P-value.
\item \code{"pathGenes"} - \code{integer} number of detected genes in each pathway.
\item \code{"geneHits"} - \code{integer} number of gene hits in each pathway.
\item \code{"itemsetID"} - \code{character} identifier for each pathway
\item \code{"Source"} - \code{character} source for each pathway, using MSigDB
subcategory where applicable.
\item \code{"Name"} - short name for each gene set or pathway.
\item \code{"Description"} - long description for each gene set of pathway.
\item \code{"Category"} - \code{character} string using MSigDB category.
\item \code{"geneNames"} - \code{character} comma-delimited gene symbols including
the number of genes in \code{"geneHits"}.
}
}
\source{
Farris et al 2019
}
\usage{
EnrichmentDFL
}
\description{
Farris et al gene set enrichment data.frames
}
\details{
This object is a \code{list} that contains one \code{data.frame}
for each gene set enrichment test.

Each gene hit list was tested against MSigDB v6.0
using hypergeometric enrichment. The MSigDB data
is available from Subramanian, Tamayo, et al.
\href{http://www.pnas.org/cgi/content/abstract/102/43/15545}{(2005, PNAS 102, 15545-15550)}
and  Liberzon, et al.
\href{https://doi.org/10.1093/bioinformatics/btr260}{(2011, Bionformatics)}
}
\examples{
data(EnrichmentDFL);

# show the first few rows of one data.frame
head(EnrichmentDFL[[1]])

# If you have the tibble package, it has a clean display
if (require(tibble)) {
   print(tibble::tibble(EnrichmentDFL[[1]]))
}

# the jamba package sdim() is a good summary
jamba::sdim(EnrichmentDFL);

}
\concept{Farris enrichment data}
\keyword{datasets}
