% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrisdata.r
\docType{package}
\name{farrisdata}
\alias{farrisdata}
\title{farrisdata: Farris et al RNA-seq data}
\description{
Supporting data for Farris et al Mouse hippocampus
RNA-seq analysis workflow, utilizing the \code{jampack}
package.
}
\details{
These objects are provided as \code{SummarizedExperiment}
objects as described in the Bioconductor package
\code{\link[SummarizedExperiment]{SummarizedExperiment}}.

All gene symbols were updated to the most recent annotations
available at the time of publication (2019) and should be updated
again for comparison with analysis results produced after
that time. One option is to use Github R package \code{"jmw86069/genejam"}
which relies upon Bioconductor gene annotation packages
such as \code{"org.Mm.eg.db"} (mouse), used with argument:
\code{ann_lib="org.Mm.eg.db"}. An example command:

\if{html}{\out{<div class="sourceCode">}}\preformatted{Ainsley_gdf <- genejam::freshenGenes3(AinsleyGenes,
   ann_lib="org.Mm.eg.db",
   intermediate="ENTREZID",
   include_source=TRUE,
   empty_rule="original")
}\if{html}{\out{</div>}}

As of October 2024, only one gene symbol was not found:
\code{"AI314180"}. All other gene symbols were correctly identified.
}
\section{Available data}{

\code{\link[farrisdata]{farrisGeneSE}},
\code{\link{farrisTxSE}}
}

\section{Gene and transcript models}{

Gene and transcript models were defined
in the corresponding Gencode GTF file
available from the Gencode FTP site:

\href{ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_mouse/release_M12/gencode.vM12.annotation.gtf.gz}{gencode.vM12.annotation.gtf.gz}

Alignment to the transcriptome was performed using
the FASTA sequence file also from Gencode FTP site:

\href{ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_mouse/release_M12/gencode.vM12.transcripts.fa.gz}{gencode.vM12.transcripts.fa.gz}

Gencode data is subject to
\href{https://creativecommons.org/licenses/by-nc-nd/2.5/}{Creative Commons License}
(Attribution, Non-commercial, No-derivatives 2.5 license),
see \url{https://creativecommons.org/licenses/by-nc-nd/2.5/}.
}

