% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrisdata.r
\docType{data}
\name{farrisTxSE}
\alias{farrisTxSE}
\title{Farris et al transcript-level expression data}
\format{SummarizedExperiment object with 122733 transcript rows, and 24
sample columns. Sample annotations are accessed as described
for \code{\link[SummarizedExperiment]{SummarizedExperiment-class}}
objects, using \code{SummarizedExperiment::colData()}. Transcript and gene
annotations are accessed using \code{SummarizedExperiment::rowData()}.

The colData colnames are described as follows:

\describe{
\item{CellType}{The major excitatory pyramidal cell types in
the mouse hippocampus, CA1, CA2, CA3, and dentate gyrus (DG).}
\item{Compartment}{The cellular compartment within each
cell type: CB (cell body) is the pyramidal cell layer;
DE (dendrite) is stratum radiatum of CA1, CA2, CA3, and the
molecular layer of DG.}
\item{AnimalID}{The number assigned to each mouse.}
\item{groupName}{The sample group name associated with each
\code{CellType} and \code{Compartment}.}
}

The count data is accessed through the function \code{assays}
and has the following entries:

\describe{
\item{tpm}{Median-normalized log2-transformed TPM values
generated by Salmon, after import using
\code{tximport::tximport()}. The log2 transformation was
log2(1+x).}
\item{counts}{Median-normalized log2-transformed pseudocount
data per-gene generated by Salmon, after import using
\code{tximport::tximport()}. The log2 transformation was
log2(1+x).}
\item{raw_tpm}{Raw log2-transformed TPM values
generated by Salmon, after import using
\code{tximport::tximport()}. The log2 transformation was
log2(1+x).}
\item{raw_counts}{The raw log2-transformed pseudocount data
per-gene generated by Salmon. The log2 transformation was
log2(1+x).}
}

The rowData contains gene annotations, which includes the relationship
between transcript and gene.

\describe{
\item{transcript_id}{Gencode transcript_id as defined in the GTF file.}
\item{geneSymbol}{Most current Entrez gene symbol associated with the
transcript_id, as of June 2018.}
\item{gene_type,transcript_type}{Gencode attributes as defined in
the GTF file.}
\item{has3UTR,TxHas3UTR}{Boolean indicating whether the gene or
transcript contained an annotated 3'UTR region defined in the
Gencode GTF file.}
\item{TxHasExt3UTR,GeneHasExt3UTR}{Following 3'UTR extension
based upon RNA-seq data coverage, this boolean field indicates whether
each transcript or gene contains a 3'UTR with or without extension
by observed RNA-seq coverage.}
\item{TxDetectedByTPM}{Boolean indicating whether each transcript
was considered "detected" based upon TPM and count data per gene,
as described in \code{splicejam::defineDetectedTx()}.}
}}
\source{
Farris et al publication.
}
\usage{
farrisTxSE
}
\description{
Farris et al transcript-level expression data
}
\details{
SummarizedExperiment object containing transcript-level expression
data following Salmon quantitation, and import using
\code{tximport::tximport()} at the transcript level.
Data is stored as log2-transformed TPM, and log2-transformed
pseudocounts from Salmon, after median normalization.
}
\examples{
data(farrisTxSE);
## Transcript-Gene annotation DataFrame
head(SummarizedExperiment::rowData(farrisTxSE))

## Sample annotation DataFrame
head(SummarizedExperiment::colData(farrisTxSE))

## Assay data matrix for normalized TPM
head(SummarizedExperiment::assays(farrisTxSE)[["tpm"]]);
}
\keyword{datasets}
