% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrisdata.r
\docType{data}
\name{GencodeVM12mm10cai}
\alias{GencodeVM12mm10cai}
\title{Codon Adaptability Index for Gencode mouse mm10 vM12}
\format{
\code{numeric} vector of codon adaptability index (cai)
values named by transcript_name defined in the Gencode
GTF file for vM12.
}
\usage{
GencodeVM12mm10cai
}
\description{
Codon Adaptability Index for Gencode mouse mm10 vM12
}
\examples{
data(GencodeVM12mm10cai);
## Print the first 20 entries
print(head(GencodeVM12mm10cai, 20));
}
\seealso{
Other Mouse codon usage data: 
\code{\link{GencodeVM12mm10caiCtBad}},
\code{\link{GencodeVM12mm10caiCtLow}},
\code{\link{GencodeVM12mm10caiQ1mean}},
\code{\link{GencodeVM12mm10cdsLength}}
}
\concept{Mouse codon usage data}
\keyword{datasets}
