% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrisdata.r
\docType{data}
\name{GencodeVM12mm10caiCtBad}
\alias{GencodeVM12mm10caiCtBad}
\title{Count of Bad CAI codons for Gencode mouse mm10 vM12}
\format{
\code{integer} number of codons where the CAI is 0.3
or lower, named by transcript_name defined in the Gencode
GTF file for vM12.
}
\usage{
GencodeVM12mm10caiCtBad
}
\description{
Count of Bad CAI codons for Gencode mouse mm10 vM12
}
\details{
This data is derived from codon adaptability index (CAI)
data, summarized to count the number of residues where
the CAI value is 0.3 or lower. This value indicates the
codon is no more than 30\% the maximum codon for a given
amino acid.
}
\examples{
data(GencodeVM12mm10caiCtBad);
## Print the first 20 entries
print(head(GencodeVM12mm10caiCtBad, 20));
}
\keyword{datasets}
