% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrisdata.r
\docType{data}
\name{GencodeVM12mm10caiQ1mean}
\alias{GencodeVM12mm10caiQ1mean}
\title{Mean of Q1 CAI values for Gencode mouse mm10 vM12}
\format{
\code{numeric} vector with the mean CAI value from the lowest
quartile of CAI values per transcript,
named by transcript_name defined in the Gencode
GTF file for vM12.
}
\usage{
GencodeVM12mm10caiQ1mean
}
\description{
Mean of Q1 CAI values for Gencode mouse mm10 vM12
}
\details{
This data is derived from codon adaptability index (CAI)
data, summarized to represent the mean of the lowest quartile
of CAI values per transcript.
}
\examples{
data(GencodeVM12mm10caiQ1mean);
## Print the first 20 entries
print(head(GencodeVM12mm10caiQ1mean, 20));
}
\seealso{
Other Mouse codon usage data: 
\code{\link{GencodeVM12mm10caiCtBad}},
\code{\link{GencodeVM12mm10caiCtLow}},
\code{\link{GencodeVM12mm10cai}},
\code{\link{GencodeVM12mm10cdsLength}}
}
\concept{Mouse codon usage data}
\keyword{datasets}
