% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrisdata.r
\docType{data}
\name{NakayamaGenes}
\alias{NakayamaGenes}
\title{Nakayama adult hippocampal CA1 dendritic genes}
\format{character vector of Entrez gene symbols}
\source{
Nakayama, K., Ohashi, R., Shinoda, Y., Yamazaki, M., Abe, M.,
Fujikawa, A., Shigenobu, S., Futatsugi, A., Noda, M., Mikoshiba, K.,
Furuichi, T., Sakimura, K., Shiina, N., 2017.
RNG105/caprin1, an RNA granule protein for dendritic mRNA localization,
is essential for long-term memory formation. eLife 6, e29677.
doi:10.7554/eLife.29677
}
\usage{
NakayamaGenes
}
\description{
Nakayama adult hippocampal CA1 dendritic genes
}
\details{
Genes derived from Nakayama et al., 2017 mouse adult hippocampal CA1
dendrites.

Gene symbols were updated using Bioconductor package
\code{org.Mm.eg.db}, using the reported \code{gene_name} or
\code{Entrez Gene ID} from the supplemental data. This step was performed
in order to ensure each reference dataset, including the Farris et al.
data, were using the same version of the gene symbol annotations.
}
\keyword{datasets}
