% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrisdata-enrich.r
\docType{data}
\name{EnrichmentDFL}
\alias{EnrichmentDFL}
\title{Farris et al gene set enrichment data.frames}
\format{An object of class \code{list} of length 12.}
\source{
Farris et al 2019
}
\usage{
EnrichmentDFL
}
\description{
Farris et al gene set enrichment data.frames
}
\details{
This object is a \code{list} that contains one \code{data.frame}
for each gene set enrichment test.

Each gene hit list was tested against MSigDB v6.0
using hypergeometric enrichment. The MSigDB data
is available from Subramanian, Tamayo, et al.
\href{http://www.pnas.org/cgi/content/abstract/102/43/15545}{(2005, PNAS 102, 15545-15550)}
and  Liberzon, et al.
\href{https://doi.org/10.1093/bioinformatics/btr260}{(2011, Bionformatics)}
}
\examples{
data(EnrichmentDFL);

# show the first few rows of one data.frame
head(EnrichmentDFL[[1]])

# If you have the tibble package, it has a clean display
if (require(tibble)) {
   print(tibble::tibble(EnrichmentDFL[[1]]))
}

# the jamba package sdim() is a good summary
jamba::sdim(EnrichmentDFL);

}
\concept{Farris enrichment data}
\keyword{datasets}
