% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrisdata.r
\docType{data}
\name{farris_sashimi_files_df}
\alias{farris_sashimi_files_df}
\title{Sashimi plot files data.frame for Farris data}
\format{data.frame suitable for use with
\code{splicejam::prepareSashimi()} for the \code{filesDF}
argument.}
\usage{
farris_sashimi_files_df
}
\description{
Sashimi plot files data.frame for Farris data
}
\examples{
#farris_sashimi_files_df
base_url <- "https://orio.niehs.nih.gov/ucscview/farrisHub/mm10/";
factor1 <- c("CA1", "CA2", "CA3", "DG");
factor2 <- c("CB", "DE");
junc_suffix <- ".STAR_mm10.combinedJunctions.bed";
junc_urls <- paste0(base_url,
   rep(factor1, each=2),
   "_",
   rep(factor2, 4),
   junc_suffix);
junc_sample_id <- paste(rep(factor1, each=2),
   rep(factor2, 4),
   sep="_");
junc_df <- data.frame(sample_id=junc_sample_id,
   url=junc_urls,
   type="junction");
if (suppressPackageStartupMessages(require(knitr))) {
   knitr::kable(junc_df)
} else {
   junc_df;
}

bw_factor1 <- c(factor1);
bw_factor2 <- c(factor2);
bw_pre1 <- c("NS50211", "NS50218", "NS50217", "NS50266", "NS50266");
bw_pre2 <- c("492", "496", "502", "492", "502");
bw_strand <- c("pos", "neg");
bw_mid <- c("_1.sickle.merged.cutadapt.STAR_mm10.",
   "-X.STAR_mm10.");
bw_suffix <- ".bw";
bw_urls <- paste0(
   rep(bw_pre1, each=16),
   rep(rep(bw_factor1, each=4), 5),
   rep(rep(rep(bw_factor2, each=2), 4), 5),
   rep(bw_pre2, each=16),
   rep(bw_mid, c(48, 32)),
   bw_strand,
   bw_suffix
);
# remove samples for which four replicates were not required
bw_urls <- gsub("^NS50266.*(CA1|CA3|496|CB).*", NA, bw_urls);
# quick view of replicates available
if (suppressPackageStartupMessages(require(knitr))) {
   knitr::kable(as.data.frame(matrix(bw_urls, ncol=5)))
} else {
   as.data.frame(matrix(bw_urls, ncol=5));
}

# remove NA entries
bw_url <- rmNA(bw_urls);
bw_sample_id <- gsub("^.*(CA[123]|DG)(CB|DE).*", "\\\\1_\\\\2",
   bw_url);
bw_df <- data.frame(
   url=paste0(base_url, bw_url),
   sample_id=bw_sample_id,
   type="bw");
bw_df <- bw_df[do.call(order, bw_df[,c(2,1,3)]),];

farris_sashimi_files_df <- rbind(
   junc_df,
   bw_df);
farris_sashimi_files_df$scale_factor <- 1;
farris_sashimi_files_df$CellType <- factor(
   gsub("_.+", "",
      farris_sashimi_files_df$sample_id),
   levels=factor1);
farris_sashimi_files_df$Compartment <- factor(
   gsub("^.+_", "",
      farris_sashimi_files_df$sample_id),
   levels=factor2);
if (suppressPackageStartupMessages(require(knitr))) {
   knitr::kable(farris_sashimi_files_df)
} else {
   farris_sashimi_files_df;
}

table(farris_sashimi_files_df[,c("CellType", "Compartment", "type")]);

}
\concept{Sashimi plot data}
\keyword{datasets}
