% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrisdata.r
\docType{package}
\name{farrisdata}
\alias{farrisdata}
\alias{farrisdata-package}
\title{farrisdata: Farris et al RNA-seq data}
\description{
Supporting data for Farris et al Mouse hippocampus
RNA-seq analysis workflow, utilizing the \code{jampack}
package.
}
\details{
These objects are provided as \code{SummarizedExperiment}
objects as described in the Bioconductor package
\code{\link[SummarizedExperiment]{SummarizedExperiment}}.
}
\section{Available data}{

   \code{\link[farrisdata]{farrisGeneSE}},
   \code{\link{farrisTxSE}}
}

\section{Gene and transcript models}{

   Gene and transcript models were defined
   in the corresponding Gencode GTF file
   available from the Gencode FTP site:

   \href{ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_mouse/release_M12/gencode.vM12.annotation.gtf.gz}{gencode.vM12.annotation.gtf.gz}

   Alignment to the transcriptome was performed using
   the FASTA sequence file also from Gencode FTP site:

   \href{ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_mouse/release_M12/gencode.vM12.transcripts.fa.gz}{gencode.vM12.transcripts.fa.gz}

   Gencode data is subject to
   \href{https://creativecommons.org/licenses/by-nc-nd/2.5/}{Creative Commons License}
   (Attribution, Non-commercial, No-derivatives 2.5 license),
   see \url{https://creativecommons.org/licenses/by-nc-nd/2.5/}.
}

