% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrisdata.r
\docType{data}
\name{GencodeVM12mm10cdsLength}
\alias{GencodeVM12mm10cdsLength}
\title{CDS length for Gencode mouse mm10 vM12}
\format{
\code{numeric} vector of CDS length for each
protein-coding transcript as defined in the Gencode
GTF file for vM12.
}
\usage{
GencodeVM12mm10cdsLength
}
\description{
CDS length for Gencode mouse mm10 vM12
}
\examples{
data(GencodeVM12mm10cdsLength);
## Print the first 20 entries
print(head(GencodeVM12mm10cdsLength, 20));
}
\seealso{
Other Mouse codon usage data: 
\code{\link{GencodeVM12mm10caiCtLow}},
\code{\link{GencodeVM12mm10caiQ1mean}}
}
\concept{Mouse codon usage data}
\keyword{datasets}
